Move `EmptyEventStore` to subpackage
