+++ server/src/main/java/io/spine/server/storage/system/SystemAwareStorageFactory.java
@@ -27,8 +27,8 @@ import io.spine.server.aggregate.Aggregate;
 import io.spine.server.aggregate.AggregateStorage;
 import io.spine.server.delivery.InboxStorage;
 import io.spine.server.entity.Entity;
-import io.spine.server.event.EmptyEventStore;
 import io.spine.server.event.EventStore;
+import io.spine.server.event.store.EmptyEventStore;
 import io.spine.server.projection.Projection;
 import io.spine.server.projection.ProjectionStorage;
 import io.spine.server.storage.RecordStorage;
